frontend: Use float64 to avoid possible int32 overflow

+++ cmd/frontend/graphqlbackend/language_statistics.go
@@ -10,8 +10,8 @@ func (l *languageStatisticsResolver) Name() string {
 	return l.l.Name
 }
 
-func (l *languageStatisticsResolver) TotalBytes() int32 {
-	return int32(l.l.TotalBytes)
+func (l *languageStatisticsResolver) TotalBytes() float64 {
+	return float64(l.l.TotalBytes)
 }
 
 func (l *languageStatisticsResolver) TotalLines() int32 {
